+++ Docs/Whats_New.md
@@ -22,6 +22,7 @@ Released TBD
 - Add correct function entry point handling.
 - Add support for `VK_KHR_get_surface_capabilities2` extension.
 - Implement newer `VK_KHR_swapchain` extension functions.
+- Add support for tracking device features enabled during `vkCreateDevice()`.
 - Allow zero offset and stride combo in `VkVertexInputBindingDescription`.
 - Fix conditions when multiple functions return VK_INCOMPLETE.
 - Fix potential memory leak on synchronous command buffer submission.